All the requirements are fullfilled. Most of the logic is written. Mainly styling is left as well as adding more quality of life features.
More testing is also needed in order to make sure no issues are present.

If you delete a project when you are on another project the to-dos will
still be deleted from the dom. In order to prevent that hide the remove
project button and only show it when the user selects a project.