Made the form easy to understand. Added a cancel button to cancel the form.
Issues Fixed:
> Users could submit the form without entering the title and description.
That is now fixed and users will be prompted to enter the title.
I have kept the description optional so it can be left empty.

>If you were using a filter and added a to-do, the to-do will show up on
the filter page. To prevent that the addToDo button will be disabled it
the user clicks on any of the filters. It will only trun back on when the
user clicks on a project.

Things to do:

> Work is required on the filter section.

> To-dos need to be stylized.

> Making a toggle for description to pop up.

> Setting the color scheme of the webpage and it's different components.

> Testing to find some other issues.